finalized the backend

+++ backend/controllers/todoControllers.js
@@ -1,48 +1,60 @@
-const Todo = require('../models/todoModel')
-const mongoose = require('mongoose')
+const Todo = require("../models/todoModel");
+const mongoose = require("mongoose");
 
-const getTodos = async(req,res) => {
-    const todos = await Todo.find()
-    res.status(200).json(todos)
-}
+const getTodos = async (req, res) => {
+  const todos = await Todo.find();
+  res.status(200).json(todos);
+};
 
-const getTodo = async(req,res) => {
-    const { id } = req.params
+const getTodo = async (req, res) => {
+  const { id } = req.params;
 
-    const todo = await Todo.findById(id)
-    res.status(200).json(todo)
+  const todo = await Todo.findById(id);
+  res.status(200).json(todo);
+};
 
-}
+const createTodo = async (req, res) => {
+  const { text } = req.body;
 
-const createTodo = async(req,res) => {
-    const { text } = req.body
-
-  try{
-    const todo = await Todo.create({text})
-    res.status(200).json(todo)
+  try {
+    const todo = await Todo.create({ text });
+    res.status(200).json(todo);
   } catch (error) {
-    res.status(400).json({error: error.message})
+    res.status(400).json({ error: error.message });
   }
-}
+};
 
-const deleteTodo = async(req,res) => {
-    const { id } = req.params
+const deleteTodo = async (req, res) => {
+  const { id } = req.params;
 
-    const todo = await Todo.findByIdAndDelete({_id: id})
-    res.status(200).json(todo)
-}
+  //check the id
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ error: "No todos" });
+  }
 
-const updateTodo = async(req,res) => {
-    const { id } =req.params
+  const todo = await Todo.findByIdAndDelete({ _id: id });
 
-    const todo = await Todo.findByIdAndUpdate({_id: id}, {
-        ...req.body
-    })
+  if (!todo) {
+    return res.status(400).json({ error: "no todos" });
+  }
+  res.status(200).json(todo);
+};
 
-    res.status(200).json(todo)
-}
+const updateTodo = async (req, res) => {
+  const { id, text } = req.body;
 
+  //check the id
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ error: "No todos" });
+  }
 
+  const todo = await Todo.findByIdAndUpdate(id, { text });
 
+  if (!todo) {
+    return res.status(400).json({ error: "no todos" });
+  }
 
-module.exports = { createTodo,getTodos,getTodo,deleteTodo,updateTodo }
+  res.status(200).json(todo);
+};
+
+module.exports = { createTodo, getTodos, getTodo, deleteTodo, updateTodo };
